requirements.txt
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-rest-registration==0.5.6
djangorestframework==3.11.0
requests==2.24.0
urllib3==1.25.10
requests-oauthlib==1.3.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-rest-auth==0.9.3
isbntools==4.3.23
